refactor: relocate estimation test data factory to common library path and update all imports
