+++ test/features/estimations/units/blocs/cost_estimation_list_bloc_test.dart
@@ -18,8 +18,8 @@ import 'package:construculator/libraries/time/testing/fake_clock_impl.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 import '../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../helpers/estimation_test_data_map_factory.dart';
 
 void main() {
   group('CostEstimationListBloc', () {
