+++ test/features/estimations/units/blocs/delete_cost_estimation_bloc_test.dart
@@ -12,7 +12,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../helpers/estimation_test_data_map_factory.dart';
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 
 void main() {
   group('DeleteCostEstimationBloc', () {
