+++ test/features/estimations/units/domain/usecases/add_cost_estimation_usecase_test.dart
@@ -16,7 +16,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../../helpers/estimation_test_data_map_factory.dart';
+import '../../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 
 void main() {
   group('AddCostEstimationUseCase', () {
