+++ test/features/estimations/widgets/accessibility/cost_estimation_landing_page_a11y_test.dart
@@ -23,7 +23,7 @@ import 'package:ripplearc_coreui/ripplearc_coreui.dart';
 
 import '../../../../utils/a11y/a11y_guidelines.dart';
 import '../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../helpers/estimation_test_data_map_factory.dart';
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 
 class _CostEstimationLandingPageA11yTestModule extends Module {
   final AppBootstrap appBootstrap;
