+++ test/features/estimations/widgets/accessibility/estimation_rename_sheet_a11y_test.dart
@@ -20,7 +20,7 @@ import 'package:ripplearc_coreui/ripplearc_coreui.dart';
 import '../../../../utils/a11y/a11y_guidelines.dart';
 import '../../../../utils/fake_app_bootstrap_factory.dart';
 import '../../../../utils/screenshot/font_loader.dart';
-import '../../helpers/estimation_test_data_map_factory.dart';
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 
 class _EstimationRenameSheetTestModule extends Module {
   final AppBootstrap appBootstrap;
