+++ test/features/estimations/widgets/pages/cost_estimation_landing_page_test.dart
@@ -35,7 +35,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:ripplearc_coreui/ripplearc_coreui.dart';
 
 import '../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../helpers/estimation_test_data_map_factory.dart';
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 
 class _CostEstimationLandingPageTestModule extends Module {
   final AppBootstrap appBootstrap;
