+++ test/features/estimations/widgets/widgets/estimation_rename_sheet_test.dart
@@ -24,7 +24,7 @@ import 'package:ripplearc_coreui/ripplearc_coreui.dart';
 
 import '../../../../utils/fake_app_bootstrap_factory.dart';
 import '../../../../utils/screenshot/font_loader.dart';
-import '../../helpers/estimation_test_data_map_factory.dart';
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart';
 
 class _EstimationRenameSheetTestModule extends Module {
   final AppBootstrap appBootstrap;
