+++ test/features/global_search/units/blocs/global_search_bloc_test.dart
@@ -2,7 +2,6 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:construculator/features/global_search/domain/entities/search_scope_entity.dart';
 import 'package:construculator/features/global_search/global_search_module.dart';
 import 'package:construculator/features/global_search/presentation/bloc/global_search_bloc/global_search_bloc.dart';
-
 import 'package:construculator/libraries/errors/failures.dart';
 import 'package:construculator/libraries/global_search/domain/search_error_type.dart';
 import 'package:construculator/libraries/supabase/data/supabase_types.dart';
@@ -14,9 +13,9 @@ import 'package:construculator/libraries/time/testing/fake_clock_impl.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../../estimations/helpers/estimation_test_data_map_factory.dart'
+import '../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart'
     as estimation_factory;
+import '../../../../utils/fake_app_bootstrap_factory.dart';
 
 // ---------------------------------------------------------------------------
 // Test data helpers
