+++ test/features/global_search/units/data/data_source/remote_global_search_data_source_test.dart
@@ -6,7 +6,6 @@ import 'package:construculator/features/global_search/data/models/pagination_par
 import 'package:construculator/features/global_search/data/models/search_params_dto.dart';
 import 'package:construculator/features/global_search/data/models/search_scope.dart';
 import 'package:construculator/features/global_search/global_search_module.dart';
-
 import 'package:construculator/libraries/supabase/data/supabase_types.dart';
 import 'package:construculator/libraries/supabase/database_constants.dart';
 import 'package:construculator/libraries/supabase/interfaces/supabase_wrapper.dart';
@@ -17,9 +16,9 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:supabase_flutter/supabase_flutter.dart' as supabase;
 
-import '../../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../../../estimations/helpers/estimation_test_data_map_factory.dart'
+import '../../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart'
     as estimation_factory;
+import '../../../../../utils/fake_app_bootstrap_factory.dart';
 
 Map<String, dynamic> _fakeSearchHistoryData({
   String? id,
