+++ test/features/global_search/units/data/repositories/global_search_repository_impl_test.dart
@@ -5,7 +5,6 @@ import 'package:construculator/features/global_search/domain/entities/search_res
 import 'package:construculator/features/global_search/domain/entities/search_scope_entity.dart';
 import 'package:construculator/features/global_search/domain/repositories/global_search_repository.dart';
 import 'package:construculator/features/global_search/global_search_module.dart';
-
 import 'package:construculator/libraries/either/either.dart';
 import 'package:construculator/libraries/errors/failures.dart';
 import 'package:construculator/libraries/global_search/domain/search_error_type.dart';
@@ -18,9 +17,9 @@ import 'package:construculator/libraries/time/testing/fake_clock_impl.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../../../utils/fake_app_bootstrap_factory.dart';
-import '../../../../estimations/helpers/estimation_test_data_map_factory.dart'
+import '../../../../../libraries/estimation/helpers/estimation_test_data_map_factory.dart'
     as estimation_factory;
+import '../../../../../utils/fake_app_bootstrap_factory.dart';
 
 // ---------------------------------------------------------------------------
 // Test data helpers
