+++ test/libraries/estimation/units/data/models/cost_estimate_dto_test.dart
@@ -5,7 +5,7 @@ import 'package:construculator/libraries/estimation/domain/entities/lock_status_
 import 'package:construculator/libraries/estimation/domain/entities/markup_configuration_entity.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../../../features/estimations/helpers/estimation_test_data_map_factory.dart';
+import '../../../helpers/estimation_test_data_map_factory.dart';
 
 void main() {
   const jsonWithNulls = {
